# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hasanaa e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: subcategoryId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: featured
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
...
  /promotions:
    get:
      operationId: listPromotions
      tags: [promotions]
      summary: List all promotion blocks
      responses:
        "200":
          description: List of promotions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promotion"
    post:
      operationId: createPromotion
      tags: [promotions]
      summary: Create a promotion block
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromotionBody"
      responses:
        "201":
          description: Promotion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
  /promotions/{id}:
    get:
      operationId: getPromotion
      tags: [promotions]
      summary: Get a promotion block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Promotion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePromotion
      tags: [promotions]
      summary: Update a promotion block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromotionBody"
      responses:
        "200":
          description: Promotion updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
    delete:
      operationId: deletePromotion
      tags: [promotions]
      summary: Delete a promotion block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Promotion deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PromotionItem:
      type: object
      properties:
        imageUrl:
          type: string
        label:
          type: string
        categoryId:
          type: ["integer", "null"]
        subcategoryId:
          type: ["integer", "null"]
      required:
        - imageUrl
        - label
    Promotion:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        gridType:
          type: integer
          description: "2 for 2-column grid, 4 for 4-column grid"
        position:
          type: string
          description: "top or bottom"
        isActive:
          type: boolean
        sortOrder:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PromotionItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - gridType
        - position
        - isActive
        - sortOrder
        - items
        - createdAt
        - updatedAt
    CreatePromotionBody:
      type: object
      properties:
        title:
          type: string
        gridType:
          type: integer
        position:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PromotionItem"
      required:
        - title
        - gridType
        - position
    UpdatePromotionBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        gridType:
          type: ["integer", "null"]
        position:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        sortOrder:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PromotionItem"
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        originalPrice:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        subcategoryId:
          type: ["integer", "null"]
        subcategoryName:
          type: ["string", "null"]
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - originalPrice
        - imageUrl
        - images
        - categoryId
        - categoryName
        - subcategoryId
        - subcategoryName
        - inStock
        - isFeatured
        - sizes
        - colors
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        originalPrice:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        subcategoryId:
          type: ["integer", "null"]
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
      required:
        - name
        - price
    UpdateProductBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        originalPrice:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: ["integer", "null"]
        subcategoryId:
          type: ["integer", "null"]
        inStock:
          type: ["boolean", "null"]
        isFeatured:
          type: ["boolean", "null"]
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
...